fix: C++11 compliance, use copy-initialization
In C++11, "auto var{ arg }" is a direct-list-initialization, which results
in the type of var being std::initializer_list, as opposed to being the
type of arg. Example: with "auto var{ 3 }" the type of var is not int.

From GCC 5.1, N3922 is implemented, which changes the above behaviour such
that the type of var is now int, even if std=c++11 is selected. However,
N3922 is officially introduced from C++17, hence breaking C++11 compliance.

The result is GCC 4.8.1 results in a build failure, but GCC 5.1 onwards
doesn't. We avoid this by properly using copy-initialization.

Signed-off-by: Adrián Herrera Arcila <adr.her.arc.95@gmail.com>
Change-Id: I21600fd88262059be50ec2e0d83a86c6a897124d